requirements.txt
appdirs==1.4.4
argcomplete==2.0.0
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
buildozer==1.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
Cython==0.29.24
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-addanother==2.0.2
django-ajax-selects==2.1.0
django-appconf==1.0.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-convenient-formsets==1.1
django-crispy-forms==1.13.0
django-dynamic-formsets==0.0.8
django-extensions==3.1.5
django-filter==2.4.0
django-formtools==2.3
django-htmx==1.6.0
django-imagekit==4.0.2
django-import-export==2.5.0
django-mathfilters==1.0.0
django-mobile-app-distribution==0.5
django-model-utils==3.1.2
django-multipleselectionfield==1.0.0
django-report-tools==0.2.2
django-search-views==0.3.7
Django-Select2==5.4.0
django-serverside-datatable==0.1.1
django-sms==0.5.0
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
docutils==0.17.1
et-xmlfile==1.1.0
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
idna==3.2
Markdown==3.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.3
pdf2image==1.16.0
pexpect==4.8.0
pilkit==2.0
Pillow==8.3.1
platformdirs==2.3.0
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.10.0
PyMuPDF==1.19.4
pyparsing==3.0.7
pyserial==3.5
python-escpos==2.2.0
pyusb==1.2.1
PyYAML==5.4.1
pyzbar==0.1.8
qrcode==7.3
reportlab==3.5.68
requests==2.26.0
sh==1.14.2
sip==6.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
thrift==0.11.0
toml==0.10.2
urllib3==1.26.6
virtualenv==20.8.0
waitress==2.0.0
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0